refactoring code!
